Move typing to type.ts (#721)
